Replaced VChanConnection with VChanInit
The previous method for initialising vchan was by adding an init
function to an otherwise standard rpc connector. As only one of the
possibly many vchan clients needs to do this initialisation, it lead to
multiple different connectors being used in a single n-1 connection.
This change separates the initialisation into a separate interface.