Upgrade linux to 4.7.2
This upgrades the bzimage to linux 4.7.2 and restructures the linux
directory to simplify upgrading linux and adding new kernel modules.
There is now a single script "build-rootfs" that compiles all kernel
modules against a particular version of linux (which it downloads if
necessary), adds the modules to the root filesystem (rootfs.cpio), and
modifies the init script to load the modules on boot.