Free interrupts on vchan module init failure
When the vmm_manager (aka vchan) module fails to load, it does not
deregister the interrupt used for the event_manager (also fails to
free some memory from kmalloc). The name of this interrupt is now
a string that was in the kernel module data segment, but once
this has been unloaded the string now points to potentially invalid
memory, resulting in a kernel crash when referenced (for instance by
doing cat /proc/interrupts).

This change adds more cleanup to the error path of the module init
to free the interrupt, as well as other resources.

The rootfs.cpio is updated to be packed with the rebuilt module