Cross vm events use existing irq inject code
Previously the interrupt was injected directly into the guest by one of
the vmm's event-handler threads. This had the potential to fail due to
the race between event handler threads and the vmm's main thread. This
change uses a notification to inform the main vmm thread of the
interrupt, and have it inject the interrupt instead.