[modules] Synthesised VFAT module.
This commit adds the VFAT FS code synthesised from Cogent, along with the
changes to `build-rootfs` script requried to add the built module to the
Linux kernel image.

Building the `VFAT` module generates a lot of warnings, which can safely be
ignored. We wnfortunately cannot fix those for now(needs changes to the Cogent
compiler).

This fixes SELFOUR-764.