Avoid short timeouts for holding OUT low
This halves the number of timeouts that need to be generated on the assumption that
no one is actually going to look specifically at the OUT signal and care about it
beind held low for a period of time. Whilst not technically a correct emulation,
in practice is not observable.