Re-port i8254 driver from qemu
Changes the i8254 emulation to be a new re-ported version from qemu. The code in qemu is
now simpler with timer changes that happened in qemu since the last version was ported.
This means that the changes for porting are less and will hopefully help remove some timing
inacuracies that current exist.