Fake hardware connector for excluding ranges of guest physical addresses
This gives a mechanism for a system assembly to state guest physical addresses that
should not be allocated for whatever reason. The motivating example is RMRR regions
defined for IOMMUs that must be mapped 1-to-1, meaning that those I/O addresses cannot
be mapped over, and so cannot be given to a guest.