cmake: Added VM Guest DMA config option
Added the ability to configure the Guests VM DMA through a CMake
configuration option. The project can toggle between DMA through
the IOMMU or no DMA. The option is exported through a configuration
library that the Init component source then includes. The IOMMU is
forced on if DMA is enabled.