vswitch: Consume used buffers after enqueuing
After enqueuing available buffers in TX we poll our send
buffqueue in case the other side has already consumed the buffer.
This is an optimisation that gives the other side a chance to
consume the outgoing traffic. This subsequently allowing us
to potentially send another packet rather than
failing because the buffer is full.