crossvm: crossvm dataports now use the VMM PCI
This change refactors the crossvm implementation to
pass dataport information to the Linux guest through a PCI device.
This change removes the need to perform on-demand mappings on
hypercall invocations. Rather the cross vm dataport support now
pre-allocates memory that gets advertised to the VM. This in
addition makes the implementation more compatible with other
platforms.