camkes: Refactor vm_config hw component into Init
This is enabled by a new camkes support for a component specifying its
own template files that aren't part of a connection/interface.

The vm_config hardware component was used to allow a VMM component to
use the template connector implementation for directly allocating
hardware resources. Allowing the vmm to use templates directly removes a
lot of change-amplification required for implementing resource
allocation through connector definitions and connector templates.