smc_cap: Allow SMC forwarding, if configured
Implemented a custom SMC call handler. It will check for a capability
associated with the SMC function ID called from the VM. If it exists,
then the SMC is forwarded to the microkernel so it can make the actual
SMC call. If it does not exist then the default SMC handler is called.
This allows PSCI emulation to continue working.

Register this new function as the VM's smc_handler_callback.

Signed-off-by: Robbie VanVossen <robert.vanvossen@dornerworks.com>