vm_arm: add option for custom device tree base
When generating a device tree, the default behavior is to use the kernel
device tree, which is passed down through the bootinfo struct. This
works for some platforms, however, there are platforms where the
bootloader or toolchain updates modify the device tree. In these cases,
the dts in the kernel is not a usable base for a generated device tree.

This commit checks if linux_image_config.dtb_base_name is in the file
server, grabs the file, and uses it as the fdt_ori pointer during the
fdt generation process.

This provides an alternative to updating the kernel device tree, keeping
that file from changing everytime a toolchain updates.

Signed-off-by: Chris Guikema <chris.guikema@dornerworks.com>