vm_arm: add allocator mempool template
The size of the allocator_mempool determines, along with the number of
untyped objects, how much RAM a VM can map. There is also a circumstance
on a 2VM system where VM0 needs to map a lot of RAM, and VM1 needs to
map a little. The current setup would require the user to increase the
allocator_mempool buffer for each VM, wasting system resources.

This commit adds a new template that defines the allocator_mempool
buffer. Simply put "vm0.allocator_mempool_size = X" in the configuration
to set the mempool size. This allows each VM to configure the mempool
for its own needs, which more efficiently utilizes system resources.

Signed-off-by: Chris Guikema <chris.guikema@dornerworks.com>